requirements.txt
numpy
opencv-python-headless
roboflow
Pillow
